+++ streamtex-pack-design/streamtex_design/cli_templates/new-gse-module/blocks/__init__.py
@@ -1,0 +1,5 @@
+"""Block registry for GSE_MODULE_NAME."""
+
+from blocks import bck_welcome
+
+__all__ = ["bck_welcome"]