+++ streamtex-pack-design/streamtex_design/cli_templates/new-gse-module/custom/themes.py
@@ -1,0 +1,5 @@
+"""Project dark-theme CSS overrides for the GSE module."""
+
+dark = {
+    "gse_brand": "color: #F5A623;",
+}